[agents] Update goap api and examples
